Handle relay-relay path
