+++ rust/relay/ebpf-turn-router/src/main.rs
@@ -16,42 +16,41 @@ mod try_handle_turn;
 #[aya_ebpf::macros::xdp]
 pub fn handle_turn(ctx: aya_ebpf::programs::XdpContext) -> u32 {
     use aya_ebpf::bindings::xdp_action;
-    use aya_log_ebpf::{debug, warn};
+    use aya_log_ebpf::{debug, trace, warn};
     use try_handle_turn::Error;
 
-    try_handle_turn::try_handle_turn(&ctx).unwrap_or_else(|e| match e {
-        Error::NotIp | Error::NotUdp => xdp_action::XDP_PASS,
+    match try_handle_turn::try_handle_turn(&ctx) {
+        Ok(()) => {
+            trace!(&ctx, target: "eBPF", "==> send packet");
 
-        Error::InterfaceIpv4AddressAccessFailed
-        | Error::InterfaceIpv6AddressAccessFailed
-        | Error::PacketTooShort
-        | Error::NotTurn
-        | Error::NoEntry(_)
-        | Error::NotAChannelDataMessage
-        | Error::UdpChecksumMissing
-        | Error::Ipv4PacketWithOptions => {
-            debug!(&ctx, "Passing packet to the stack: {}", e);
+            xdp_action::XDP_TX
+        }
+        Err(Error::NotIp | Error::NotUdp) => xdp_action::XDP_PASS,
+        Err(
+            e @ (Error::InterfaceIpv4AddressAccessFailed
+            | Error::InterfaceIpv6AddressAccessFailed
+            | Error::PacketTooShort
+            | Error::NotTurn
+            | Error::NotAChannelDataMessage
+            | Error::UdpChecksumMissing
+            | Error::Ipv4PacketWithOptions),
+        ) => {
+            debug!(&ctx, target: "eBPF", "^^^ pass packet to userspace: {}", e);
 
             xdp_action::XDP_PASS
         }
-
-        // TODO: Remove this when same-host relay-relay is supported.
-        Error::PacketLoop => {
-            debug!(&ctx, "Dropping packet: {}", e);
-
-            xdp_action::XDP_DROP
-        }
-
-        // These are exceptions and shouldn't happen in practice - WARN.
-        Error::BadChannelDataLength
-        | Error::InterfaceIpv4AddressNotConfigured
-        | Error::InterfaceIpv6AddressNotConfigured
-        | Error::XdpAdjustHeadFailed(_) => {
-            warn!(&ctx, "Dropping packet: {}", e);
+        Err(
+            e @ (Error::InterfaceIpv4AddressNotConfigured
+            | Error::InterfaceIpv6AddressNotConfigured
+            | Error::NoEntry(_)
+            | Error::BadChannelDataLength
+            | Error::XdpAdjustHeadFailed(_)),
+        ) => {
+            warn!(&ctx,target: "eBPF", "XXX drop packet: {}", e);
 
             xdp_action::XDP_DROP
         }
-    })
+    }
 }
 
 /// Defines our panic handler.
