+++ rust/relay/ebpf-turn-router/src/try_handle_turn/channel_data.rs
@@ -7,4 +7,12 @@ pub struct CdHdr {
 
 impl CdHdr {
     pub const LEN: usize = core::mem::size_of::<Self>();
+
+    pub fn number(&self) -> u16 {
+        u16::from_be_bytes(self.number)
+    }
+
+    pub fn length(&self) -> u16 {
+        u16::from_be_bytes(self.length)
+    }
 }
